Refactor ZIP download functionality and integrate into MainContentHeader component
- Moved ZIP download logic to a new custom hook `useDownloadZip`.
- Updated `MainContentHeader` to utilize the new hook for downloading package files.
- Removed redundant code related to package file fetching from `MainContentHeader`.
- Passed `packageFiles` as a prop to `MainContentHeader` from `RepoPageContent`.
